[Silver IV] Title: 스케이트 연습, Time: 72 ms, Memory: 9836 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/28324. 스케이트 연습/스케이트 연습.cc
@@ -1,6 +1,5 @@
 #include <iostream>
 #include <vector>
-#include <algorithm>
 using namespace std;
 
 int main(void)
@@ -10,8 +9,8 @@ int main(void)
 
 	int N;
 	cin >> N;
-	vector<int> vec(N);
-	vector<int> vec2(N);
+	vector<long long> vec(N);
+	vector<long long> vec2(N);
 
 	for (int i = N - 1; i >= 0; i--)
 	{
@@ -19,7 +18,7 @@ int main(void)
 	}
 
 	vec2[0] = 1;
-	int sum = 1;
+	long long sum = 1;
 	for (int i = 1; i < N; i++)
 	{
 		if (vec[i] >= vec2[i - 1] + 1)
